+++ plot_data/templates.py
@@ -110,7 +110,7 @@ histogram_template = Template('''
             var height = Math.max(0.95*window.innerHeight, 350);
 
             var data = $data;
-            var plot_data = new PlotData.newHistogram(
+            var plot_data = new PlotData.Histogram(
                 data, width, height, true, 0, 0, $canvas_id.id
             );
             plot_data.define_canvas($canvas_id.id);