+++ src/subplots.ts
@@ -1846,7 +1846,7 @@ export class Frame extends BasePlot {
   }
 }
 
-export class newHistogram extends Frame {
+export class Histogram extends Frame {
   public bars: Bar[] = [];
   readonly barsColorFill: string = 'hsl(203, 90%, 85%)';
   readonly barsColorStroke: string = 'hsl(0, 0%, 0%)';